Thread filtering using the stack aggregation view
Thread filtering is now enabled using the stack aggregation view.
Multiple threads can be selected at the same time and a button has
been added to send the filter to GDB.

The next step would be to enable thread selection from the stack view
to be linked with thread selection from the main debug view.